requirements.txt
Django~=5.1.1
djangorestframework==3.12.4
django-cors-headers==3.10.0
psycopg2-binary==2.9.1
django-allauth==0.45.0
channels==3.0.4
channels-redis==3.3.1